ep09-core-pages/bower_components/web-animations-next/test/blink/get-css-players.html

<!--
 Copyright 2017 Google Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
      http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-->

<!DOCTYPE html>
<script src="testharness/testharness.js"></script>
<script src="testharness/testharnessreport.js"></script>
<style>
@keyframes test {
    from { opacity: 0; }
    to { opacity: 1; }
}
.cssAnimation {
    animation: test 2s;
}
</style>
<div id='container'>
    <div id='element'></div>
</div>

<script>
async_test(function(t) {
    assert_equals(document.timeline.getAnimationPlayers().length, 0);
    assert_equals(container.getAnimationPlayers().length, 0);
    assert_equals(element.getAnimationPlayers().length, 0);

    element.className = 'cssAnimation';
    onload = function () {
        t.step(function() {
            var players = document.timeline.getAnimationPlayers();
            assert_equals(players.length, 1);
            assert_equals(container.getAnimationPlayers().length, 0);
            assert_equals(element.getAnimationPlayers().length, 1);

            players[0].finish();
            assert_equals(document.timeline.getAnimationPlayers().length, 0);
            assert_equals(container.getAnimationPlayers().length, 0);
            assert_equals(element.getAnimationPlayers().length, 0);
            t.done();
        });
    }
}, 'getAnimationPlayers() with cssanimations');

</script>
